requirements.txt
customtkinter
google-generativeai